libvsmsocket: add vsm_socket_select() to wait for client or data
Add a convenience function vsm_socket_select() to wait for an incoming
client connection or some input data to read.  This is useful in a
multi-threaded environment, for example with one thread to read and
one to write, as this function can be called without any lock to not
block the other thread.

Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>